Accelerate BlirRectNoMaskComplete
This was a tricky one. When implemented, P96 replaces many (all?) of the BlitRect (`rect_copy`) that use a mask of 0xFF to allow for graphics minterm handling instead.
I suspect this is due to pretty much running out of registers when calling both of these functions.
The minterm handling currently isn't implemented, but it seems to work fine for all the test cases I have available.